+++ CHANGES.rst
@@ -2,6 +2,7 @@
 =======================
 
 - #53: Fix source distribution
+- #54: Fix typo in test
 
 
 0.16 (Jul 25, 2012)